move Graph.h to its own directory
