+++ arangod/Pregel/Algos/EffectiveCloseness/HLLCounter.cpp
@@ -24,7 +24,7 @@
 #include <cmath>
 #include "Basics/fasthash.h"
 #include "Pregel/CommonFormats.h"
-#include "Pregel/Graph.h"
+#include "Pregel/Graph/Graph.h"
 
 using namespace arangodb::pregel;
