+++ arangod/Pregel/VertexComputation.h
@@ -26,8 +26,9 @@
 #include <algorithm>
 #include <cstddef>
 #include "Basics/Common.h"
-#include "Pregel/Graph.h"
+#include "Pregel/Graph/Graph.h"
 #include "Pregel/Worker/GraphStore.h"
+#include "Pregel/Graph/Graph.h"
 #include "Pregel/OutgoingCache.h"
 #include "Pregel/Worker/WorkerConfig.h"
 #include "Pregel/WorkerContext.h"
